+++ oneflow/user/ops/copy_op.cpp
@@ -35,7 +35,7 @@ std::function<Maybe<const Device>(user_op::DeviceInferContext* ctx)> GetDeviceIn
   std::function<Maybe<const Device>(user_op::DeviceInferContext * ctx)> fn =
       [](user_op::DeviceInferContext* ctx) -> Maybe<const Device> {
     std::shared_ptr<const Device> out_device =
-        JUST(Device::New(ctx->Attr<std::string>("device_type"), ctx->Attr<int>("device_id")));
+        JUST(Device::New(ctx->Attr<std::string>("device_type"), ctx->Attr<int64_t>("device_id")));
     *ctx->OutputTensorDevice4ArgNameAndIndex("out", 0) = out_device;
     const std::shared_ptr<const Device>& in_device =
         ctx->InputTensorDevice4ArgNameAndIndex("in", 0);
